internal/scripts/requirements.txt
torch
torchvision
torchaudio
transformers
requests
onnx
onnxruntime
Pillow
numpy